requirements.txt
# Web scraping
browserbase>=0.1.0
playwright>=1.40.0

# Vector database
qdrant-client>=1.7.0

# Embeddings (choose one or both)
openai>=1.0.0  # Optional: for OpenAI embeddings
sentence-transformers>=2.2.0  # Free local embeddings (recommended)
tf-keras  # Keras compatibility for sentence-transformers

# Data processing
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
tqdm>=4.66.0

# Web UI
flask>=3.0.0
flask-cors>=4.0.0

# LLM for repo selection
groq>=0.4.0
